Crankshaft[BU000001WFSKM9] Client - Ruby @ 2015-08-21 14:26:49 +0000
57b0851 Jack Franklin <jack@jackfranklin.net>
Merge pull request #28 from gocardless/fix-links-bug

Fix ordering of links on resources

----------------

4fa473e Jack Franklin <jack@jackfranklin.net>
Fix ordering of links on resources

We were doing some clever sorting of links to figure out how to
assign them in the right order to the struct, this falls down when we
deal with resources like events that may or may not have certain links.
Instead of trying to rely on some ordering, instead we just override the
init of the struct to take a hash, and loop through all keys in there.

----------------