Allow handling itempotency conflicts using exceptions
Integrators cannot currently handle idempotency conflicts, because our library
swallows the exception and simply refetches the resource.

This is normally a reasonable behaviour, however this means that you also can't
detect the failure and do something useful.

This commit allows configuration for the client, to explicitly re-raise an
exception to be handled appropriately.