Support both Faraday 0.x and 1.x
There was a small breaking change about passing a `nil` value for
connection opts, which is resolved by having a hash default.

To ensure compatibility with both versions of Faraday, we run the specs
parameterised with 0.9.2 and ~> 1.0 (these can be updated later). Note
that the 1.x series is not compatible with Ruby <2.3 so those are
explicitly excluded from the matrix in CI.