Fix Ruby 2.7 URI.escape deprecation warning
Switching to using CGI.escape accomplishes a similar effect. Although
they behave slightly differently, in this case it shouldn't be a problem
since we only ever use this function for replacing `:identity` with
something like `PM123`. If you send us any garbage in that string, we'll
uri-encode it so that it's no longer valid anyway and the server will
404.

There's nothing stopping us from sending some other parameter than
:identity in future however so I chose not to move to a regular
whitelist (`[A-z0-9]`).